requirements.txt
asgiref==3.8.1
Django==4.2.16
django-cuser==2017.3.16
django-modeltranslation==0.19.9
django-tinymce==4.1.0
pillow==10.4.0
sqlparse==0.5.1
typing_extensions==4.12.2
django-admin-sortable2==2.2.3
